+++ frontend/src/app/pages/cliente/solicitacao/visualizar-solicitacao/visualizar-solicitacao.component.ts
@@ -2,9 +2,7 @@ import { Component, OnInit, TemplateRef, ViewChild } from "@angular/core";
 import { FormsModule, NgForm } from "@angular/forms";
 import { Solicitacao, Equipamento,Orcamento,EstadosSolicitacao } from "../../../../shared/models";
 import { ActivatedRoute, RouterModule } from "@angular/router";
-import { SolicitacaoService } from "../../../../services/solicitacao.service";
-import { EquipamentoService } from "../../../../services/equipamento.service";
-import { OrcamentoService } from "../../../../services/orcamento.service";
+import { SolicitacaoService,EquipamentoService,OrcamentoService } from "../../../../services";
 import { CommonModule } from "@angular/common";
 import { ModalComponent } from "../../../../components/modal/modal.component";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
