+++ frontend/src/app/pages/empregado/funcionario/inserir-funcionario/inserir-funcionario.component.ts
@@ -1,7 +1,7 @@
 import { Component, ViewChild } from "@angular/core";
 import { FormsModule, NgForm } from "@angular/forms";
 import { Funcionario } from "../../../../shared/models";
-import { FuncionarioService } from "../../../../services/funcionario.service";
+import { FuncionarioService } from "../../../../services";
 import { Router, RouterModule } from "@angular/router";
 import { CommonModule } from "@angular/common";
