+++ frontend/src/app/pages/empregado/solicitacao/visualizar-solicitacao/visualizar-solicitacao.component.ts
@@ -2,13 +2,8 @@ import { Component, OnInit, TemplateRef, ViewChild } from "@angular/core";
 import { FormsModule, NgForm } from "@angular/forms";
 import { Cliente, Equipamento,Funcionario, Solicitacao, Orcamento,EstadosSolicitacao } from "../../../../shared/models";
 import { ActivatedRoute, Router, RouterModule } from "@angular/router";
-import { SolicitacaoService } from "../../../../services/solicitacao.service";
-import { EquipamentoService } from "../../../../services/equipamento.service";
-import { FuncionarioService } from "../../../../services/funcionario.service";
-import { OrcamentoService } from "../../../../services/orcamento.service";
-import { ClienteService } from "../../../../services/cliente.service";
+import { SolicitacaoService,EquipamentoService,FuncionarioService,OrcamentoService,ClienteService,LoginService } from "../../../../services";
 import { CommonModule } from "@angular/common";
-import { LoginService } from "../../../../services/login/login.service";
 import { ModalComponent } from "../../../../components/modal/modal.component";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
 import { EstadoAmigavelPipe } from "../../../../shared/pipes/estado-amigavel.pipe";
