+++ frontend/src/app/services/cep/index.ts
@@ -1,0 +1,1 @@
+export * from './cep.service'
