+++ frontend/src/app/services/index.ts
@@ -1,0 +1,9 @@
+export * from './cep';
+export * from './login';
+export * from './service-crud-base';
+export * from './cliente.service';
+export * from './equipamento.service';
+export * from './funcionario.service';
+export * from './id-base.service';
+export * from './orcamento.service';
+export * from './solicitacao.service';
