+++ frontend/src/app/services/service-crud-base/index.ts
@@ -1,0 +1,1 @@
+export * from './service-crud-base'
